+++ lib/loofah.rb
@@ -5,7 +5,7 @@ require 'nokogiri'
 require 'loofah/metahelpers'
 require 'loofah/elements'
 
-require 'loofah/html5/whitelist'
+require 'loofah/html5/safelist'
 require 'loofah/html5/libxml2_workarounds'
 require 'loofah/html5/scrub'